circuit-macros: Move tests to `circuits` crate
This is done to avoid duplicating abstractions between the two
crates. A proc-macro crate cannot export items that are not proc-macros,
so we define the necessary abstractions once in the `circuits` crate
and use them for both testing and production.
